refactor(chat): extract artist selection + chat-session provisioning per OCP
Addresses sweetmantech's review comments on lines `useArtists.tsx:80`
and `useNewChatBootstrap.ts:107` — net-new logic was being added inline
to existing hooks instead of being given its own module.

Two extractions:

1. `hooks/artists/useArtistSelection.ts` — owns the per-org artist
   selection: `useLocalStorage` for saved picks, `userOverride` state
   for explicit this-session pick/deselect, and the derived
   `selectedArtist` memo (precedence: override → saved → first).
   `useArtists.tsx` now just calls it with `(orgKey, artists)`.

2. `lib/sessions/provisionChatSession.ts` — combines `createSession`
   + `createSandbox` into one function. `useNewChatBootstrap.ts` now
   wraps that in `useMutation` and owns the trigger / input-change
   re-fire logic only.

No behavior change — both hooks return the same shape and respond to
the same inputs. Just smaller files with clearer single responsibility.
